Remove detection by default
Previously, even without a class on `code`, its programming language
was inferred, and it was highlighted anyway.
It was possible to turn this by passing `subset: false`, an overload
of an option that also has another job.

This commit reverts this behavior, and removes the overload in favor of
a new option: pass `detect: true` to automatically detect the
programming language of code blocks.
The default, `detect: false`, does not do that.

If you had `subset: false`, you can remove that line.
If you want the previous behavior, pass `detect: true`.

Closes GH-21.